@@ -1,12 +1,13 @@
 source 'https://rubygems.org'
 
+ruby '>= 3.1.6'
+
 gemspec
-gem 'rails', '~> 6.1.0'
+gem 'rails', '~> 7.0.0'
 gem 'pg'
 gem 'active_model_serializers'
 
 group :development, :test do
-  gem 'phony'
   gem 'faker'
   gem 'dotenv-rails'
   gem 'byebug', platforms: [:mri, :mingw, :x64_mingw]
@@ -15,16 +16,24 @@ group :development, :test do
   gem 'database_cleaner-active_record'
   gem 'database_cleaner-redis'
   gem 'factory_bot_rails'
-  gem 'rubocop', require: false
+  gem 'rubocop'
+  gem 'rubocop-factory_bot'
+  gem 'rubocop-performance'
+  gem 'rubocop-rails'
+  gem 'rubocop-rspec'
+  gem 'rubocop-rspec_rails'
   gem 'pry'
   gem 'pry-rails'
   gem 'pry-byebug'
   gem 'spring-commands-rspec'
 
   # Identity requirements
-  gem 'sidekiq'
-  gem 'sidekiq-batch'
-  gem 'sidekiq-limit_fetch'
-  gem 'sidekiq-unique-jobs'
+  gem 'audited', '~> 5.4.2'
+  gem 'phony'
+  gem 'redis', '~> 4.8'
+  gem 'sidekiq', '~> 6.5.10'
+  gem 'sidekiq-batch', '~> 0.1.8'
+  gem 'sidekiq-killswitch'
+  gem 'sidekiq-unique-jobs', '~> 7.1.33'
   gem 'zip'
 end
