@@ -1,8 +1,6 @@
 describe IdentitySpoke::CampaignContact do
   context '#add_members' do
     before(:each) do
-      clean_external_database
-
       @spoke_organization = FactoryBot.create(:spoke_organization)
       @spoke_campaign = FactoryBot.create(:spoke_campaign, organization: @spoke_organization)
       @member = FactoryBot.create(:member_with_mobile)
@@ -18,11 +16,11 @@ describe IdentitySpoke::CampaignContact do
     it 'has inserted the correct campaign contacts to Spoke' do
       IdentitySpoke::CampaignContact.add_members(@spoke_campaign.id, @rows)
       expect(@spoke_campaign.campaign_contacts.count).to eq(2)
-      expect(@spoke_campaign.campaign_contacts.find_by_cell("+#{@member.phone_numbers.mobile.first.phone}").first_name).to eq(@member.first_name) # Spoke allows external IDs to be text
+      expect(@spoke_campaign.campaign_contacts.find_by(cell: "+#{@member.phone_numbers.mobile.first.phone}").first_name).to eq(@member.first_name) # Spoke allows external IDs to be text
     end
 
     it "doesn't insert duplicates already existing into Spoke" do
-      2.times do |index|
+      2.times do |_index|
         IdentitySpoke::CampaignContact.add_members(@spoke_campaign.id, @rows)
       end
       expect(@spoke_campaign.campaign_contacts.count).to eq(2)
@@ -31,7 +29,7 @@ describe IdentitySpoke::CampaignContact do
 
     it "doesn't insert duplicates from the same batch into Spoke" do
       double_up = ActiveModel::Serializer::CollectionSerializer.new(
-        [@member,@member],
+        [@member, @member],
         serializer: IdentitySpoke::SpokeMemberSyncPushSerializer,
         campaign_id: @spoke_campaign.id
       ).as_json
