@@ -1,7 +1,7 @@
 module IdentitySpoke
   module ApplicationHelper
     def self.campaigns_for_select
-      IdentitySpoke::Campaign.all.order("id DESC").map { |x| ["#{x.id}: #{x.title}", x.id] }
+      IdentitySpoke::Campaign.order("id DESC").map { |x| ["#{x.id}: #{x.title}", x.id] }
     end
   end
 end
