@@ -2,8 +2,9 @@ module IdentitySpoke
   module ConnectionExtension
     def self.included(base)
       base.class_eval do
-        def self.bulk_create(set=[], conflict=false)
+        def self.bulk_create(set = [], _conflict = false)
           return if set.empty?
+
           field_keys = *set[0].keys
           values_string = set.map do |x|
             values = field_keys.map do |field_key|
