@@ -4,7 +4,7 @@ module IdentitySpoke
     self.table_name = "campaign_contact"
     belongs_to :campaign
     belongs_to :assignment, optional: true
-    has_many :question_responses
+    has_many :question_responses, dependent: nil
 
     def self.add_members(campaign_id, batch)
       OptOut.transaction do
@@ -20,24 +20,21 @@ module IdentitySpoke
       end
     end
 
-    private
-
     def self.remove_exclusions_from_batch(campaign_id, batch)
       # Remove any duplicates in the same batch
       batch = batch.uniq { |row| row[:cell] }
 
-      cells = batch.map{ |row| row[:cell] }
+      cells = batch.pluck(:cell)
 
       # Remove any already-opted-out phone numbers
-      opted_out_cells = OptOut.where('cell in (?)', cells ).map{|opt_out| opt_out[:cell] }
-      batch = batch.reject{|row| opted_out_cells.include?(row[:cell]) }
+      opted_out_cells = OptOut.where(cell: cells).pluck(:cell)
+      batch = batch.reject { |row| opted_out_cells.include?(row[:cell]) }
 
       # Remove any phone numbers already in the campaign
-      existing_cells = where('cell in (?)', cells)
-                         .where('campaign_id = ?', campaign_id)
-                         .map{|contact| contact[:cell] }
-      batch.reject{|row| existing_cells.include?(row[:cell]) }
+      existing_cells = where(cell: cells)
+                       .where(campaign_id: campaign_id)
+                       .pluck(:cell)
+      batch.reject { |row| existing_cells.include?(row[:cell]) }
     end
-
   end
 end
