@@ -2,6 +2,6 @@ module IdentitySpoke
   class InteractionStep < ReadOnly
     self.table_name = "interaction_step"
     belongs_to :campaign
-    has_many :question_responses
+    has_many :question_responses, dependent: nil
   end
 end
