@@ -1,7 +1,7 @@
 module IdentitySpoke
   class Organization < ReadOnly
     self.table_name = "organization"
-    has_many :campaigns
-    has_many :opt_outs
+    has_many :campaigns, dependent: nil
+    has_many :opt_outs, dependent: nil
   end
 end
