@@ -1,7 +1,7 @@
 module IdentitySpoke
   class User < ReadOnly
     self.table_name = "user"
-    has_many :assignment
-    has_many :message
+    has_many :assignment, dependent: nil
+    has_many :message, dependent: nil
   end
 end
