@@ -1,3 +1,3 @@
 module IdentitySpoke
-  VERSION = '3.0.0'
+  VERSION = '3.0.0'.freeze
 end
