@@ -1,5 +1,7 @@
-
 //= link_tree ../images
 //= link_directory ../javascripts .js
 //= link_directory ../stylesheets .css
-//= link identity_nation_builder_manifest.js
+
+//= link application.js
+//= link application.css
+//= link_tree ../builds
