@@ -12,9 +12,11 @@
 #
 
 class Area < ApplicationRecord
+  audited
 
   has_many :area_zips
-  has_and_belongs_to_many :members, join_table: :area_memberships
+  has_many :area_memberships
+  has_many :members, through: :area_memberships
   has_and_belongs_to_many :canonical_addresses
 
   validates_uniqueness_of :code, scope: :area_type
@@ -56,7 +58,7 @@ class Area < ApplicationRecord
   end
 
   def icon
-    Area.icons[area_type]
+    area_icon || Area.icons[area_type]
   end
 
   def friendly_name
