@@ -1,7 +1,5 @@
 class ContactResponse < ApplicationRecord
-  include ReadWriteIdentity
-  attr_accessor :audit_data
-    
   belongs_to :contact
   belongs_to :contact_response_key
+  validates_presence_of :contact
 end
