@@ -1,7 +1,4 @@
 class ContactResponseKey < ApplicationRecord
-  include ReadWriteIdentity
-  attr_accessor :audit_data
-    
-  belongs_to :contact_campaign
+  belongs_to :contact_campaign, optional: true
   has_many :contact_responses
 end
